Change serialize and deserialize functions to return error code and fix part of adaptive timeout algorithm

+++ src/include/adaptive_delay_model.h
@@ -22,7 +22,7 @@ typedef struct adaptive_timeout_struct {
     double alpha;
     double var;
     double error;
-    long seq_num;
+    long long seq_num;
     double sum_w1;
     double sum_w2;
     double sum_time_deltas;
